stdio: Add casts to make signed/unsigned compares explicit
stdio uses 'int' for various size values in the FILE struct, which
makes gcc generate warnings when comparing against actual size
information. Add explicit casts.

+++ newlib/libc/stdio/fgets.c
@@ -162,8 +162,8 @@ _fgets_r (struct _reent * ptr,
        * newline, and stop.  Otherwise, copy entire chunk
        * and loop.
        */
-      if (len > n)
-	len = n;
+      if (len > (size_t) n)     /* n is always non-negative here */
+        len = (size_t) n;
       t = (unsigned char *) memchr ((void *) p, '\n', len);
       if (t != 0)
 	{
